+++ editorextensions/codebeamer-custom-ui/modal/src/api/lucidGateway.ts
@@ -3,133 +3,9 @@ import getItemColorField from './utils/getItemColorField';
 import { store } from '../store/store';
 import TrackerDetails from '../models/trackerDetails.if';
 import { CardData } from '../models/lucidCardData';
-
-/**
- * Class for handling message events and callbacks.
- */
-export class MessageHandler {
-	private callbacks: ((data: []) => void)[] = [];
-
-	/**
-	 * Private instance to hold the singleton instance.
-	 * @type {MessageHandler | null}
-	 * @private
-	 */
-	private static instance: MessageHandler | null = null;
-
-	/**
-	 * Private constructor to prevent direct instantiation. Sets up a message event listener.
-	 * @private
-	 */
-	private constructor() {
-		window.addEventListener('message', (e) => {
-			const data = JSON.parse(e.data);
-			this.notifyCallbacks(data);
-		});
-	}
-
-	/**
-	 * Gets the singleton instance of the MessageHandler.
-	 * If the instance doesn't exist, it creates one.
-	 * @returns {MessageHandler} The singleton instance of MessageHandler.
-	 * @static
-	 */
-	public static getInstance(): MessageHandler {
-		if (!MessageHandler.instance) {
-			MessageHandler.instance = new MessageHandler();
-		}
-		return MessageHandler.instance;
-	}
-
-	/**
-	 * Requests card blocks from the parent window and registers a callback to handle the response.
-	 * @param {function} callback - The callback function that will be called with the received card block data.
-	 */
-	getCardBlocks(callback: (data: CardBlockData[]) => void) {
-		this.subscribeCallback(callback);
-
-		LucidGateway.requestCardBlockData();
-	}
-
-	/**
-	 * Register a callback function to handle messages.
-	 * @param callback - The callback function to register.
-	 */
-	subscribeCallback(callback: (data: []) => void) {
-		this.callbacks.push(callback);
-	}
-
-	/**
-	 * Unregister a previously registered callback function.
-	 * @param callback - The callback function to unregister.
-	 */
-	unsubscribeCallback(callback: (data: []) => void) {
-		const index = this.callbacks.indexOf(callback);
-		if (index !== -1) {
-			this.callbacks.splice(index, 1);
-		}
-	}
-
-	/**
-	 * Notifies registered callbacks with message data.
-	 * @param data - The data received in the message.
-	 */
-	private notifyCallbacks(data: []) {
-		this.callbacks.forEach((callback) => callback(data));
-	}
-}
-
-/**
- * Interface for a message object with an action and payload.
- * @interface Message
- * @property action - The action to perform.
- * @property payload - The payload for the action.
- */
-export interface Message {
-	action: MessageAction;
-	payload?: ImportItemPayload | UpdateCardPayload | StartImportPayload;
-}
-
-// Interface for payload specific to the IMPORT_ITEM action
-export interface ImportItemPayload {
-	importId: number;
-	cardData: CardData;
-}
-
-// Interface for payload specific to the UPDATE_CARD action
-export interface UpdateCardPayload {
-	cardData: CardData;
-	cardBlockId: string;
-}
-
-// Interface for payload specific to the START_IMPORT action
-export interface StartImportPayload {
-	id: number;
-	totalItems: number;
-}
-
-/**
- * action enum for message actions
- */
-export enum MessageAction {
-	GET_CARD_BLOCKS = 'getCardBlocks',
-	IMPORT_ITEM = 'importItem',
-	UPDATE_CARD = 'updateCard',
-	CREATE_CONNECTORS = 'createConnectors',
-	START_IMPORT = 'startImport',
-	CLOSE_MODAL = 'closeModal',
-}
-
-/**
- * Interface for card block data.
- */
-export interface CardBlockData {
-	cardBlock: {
-		id: string;
-	};
-	codebeamerItemId: number;
-	codebeamerTrackerId: number;
-}
+import { RelationshipType } from '../enums/associationRelationshipType.enum';
+import { getColorForRelationshipType } from './utils/getColorForRelationshipType';
+import { Message, MessageAction } from '../models/messageInterfaces';
 
 export class LucidGateway {
 	/**
@@ -166,8 +42,31 @@ export class LucidGateway {
 		});
 	}
 
-	public static async createConnectors() {
-		throw new Error('Not implemented');
+	/**
+	 * Create a line between two card blocks
+	 * @param importId - The ID of the import.
+	 * @param sourceBlockId - The ID of the source card block.
+	 * @param targetBlockId - The ID of the target card block.
+	 * @param relationshipType - The type of relationship between the codebeamer items on the two card blocks.
+	 */
+	public static async createLine(
+		importId: number,
+		sourceBlockId: string,
+		targetBlockId: string,
+		relationshipType: RelationshipType
+	) {
+		const lineColor = getColorForRelationshipType(relationshipType);
+
+		this.postMessage({
+			action: MessageAction.CREATE_LINE,
+			payload: {
+				importId,
+				sourceBlockId,
+				targetBlockId,
+				relationshipType,
+				lineColor,
+			},
+		});
 	}
 
 	/**
@@ -271,6 +170,27 @@ export class LucidGateway {
 	public static requestCardBlockData() {
 		this.postMessage({
 			action: MessageAction.GET_CARD_BLOCKS,
+		});
+	}
+
+	/**
+	 * Request line data from the parent window.
+	 */
+	public static requestLineData() {
+		this.postMessage({
+			action: MessageAction.GET_LINES,
+		});
+	}
+
+	/**
+	 * Delete a line by id
+	 * @param importId - The ID of the import.
+	 * @param lineId - The ID of the line to delete.
+	 */
+	public static deleteLine(importId: number, lineId: string) {
+		this.postMessage({
+			action: MessageAction.DELETE_LINE,
+			payload: { importId, lineId },
 		});
 	}