+++ editorextensions/codebeamer-custom-ui/modal/src/pages/import/components/importer/importer.cy.tsx
@@ -2,28 +2,31 @@ import * as React from 'react';
 import { setTrackerId } from '../../../../store/slices/userSettingsSlice';
 import { getStore } from '../../../../store/store';
 import Importer from './Importer';
+import { CardBlockToCodebeamerItemMapping } from '../../../../models/lucidCardData';
+import { BlockRelation, LucidLineData } from '../../../../models/lucidLineData';
+import { RelationshipType } from '../../../../enums/associationRelationshipType.enum';
 
 describe('<Importer>', () => {
 	it('mounts', () => {
-		cy.mountWithStore(<Importer items={[]} />);
+		cy.mountWithStore(<Importer items={[]} mode="" />);
 	});
 
 	it('does not show a close button by default', () => {
-		cy.mountWithStore(<Importer items={[]} />);
+		cy.mountWithStore(<Importer items={[]} mode="" />);
 
 		cy.get('[aria-label="close"]').should('not.exist');
 	});
 
 	it('does show a close button if passed an onClose prop', () => {
 		const handler = cy.spy();
-		cy.mountWithStore(<Importer items={[]} onClose={handler} />);
+		cy.mountWithStore(<Importer items={[]} mode="" onClose={handler} />);
 
 		cy.get('[aria-label="Close"]').should('exist');
 	});
 
 	it('calls the passed onClose handler when the close button is clicked', () => {
 		const handler = cy.spy().as('handler');
-		cy.mountWithStore(<Importer items={[]} onClose={handler} />);
+		cy.mountWithStore(<Importer items={[]} mode="" onClose={handler} />);
 
 		cy.get('[aria-label="Close"]').click();
 
@@ -39,7 +42,9 @@ describe('<Importer>', () => {
 
 		cy.intercept('POST', '**/api/v3/items/query').as('fetch');
 
-		cy.mountWithStore(<Importer items={items} />, { reduxStore: store });
+		cy.mountWithStore(<Importer items={items} mode="import" />, {
+			reduxStore: store,
+		});
 
 		cy.wait('@fetch')
 			.its('request.body.queryString')
@@ -61,7 +66,9 @@ describe('<Importer>', () => {
 		cy.intercept('GET', '**/api/v3/trackers/103').as('trackerFetchThree');
 		cy.intercept('GET', '**/api/v3/trackers/104').as('trackerFetchFour');
 
-		cy.mountWithStore(<Importer items={items} />, { reduxStore: store });
+		cy.mountWithStore(<Importer items={items} mode="import" />, {
+			reduxStore: store,
+		});
 
 		cy.wait('@fetch');
 
@@ -91,7 +98,9 @@ describe('<Importer>', () => {
 
 		cy.intercept('POST', '**/api/v3/items/query').as('fetch');
 
-		cy.mountWithStore(<Importer items={items} />, { reduxStore: store });
+		cy.mountWithStore(<Importer items={items} mode="import" />, {
+			reduxStore: store,
+		});
 
 		cy.wait('@fetch')
 			.its('request.body.queryString')
@@ -102,7 +111,13 @@ describe('<Importer>', () => {
 		cy.stub(window.parent, 'postMessage')
 			.as('boardGetStub')
 			.callsFake(() => {
-				window.postMessage(JSON.stringify(mockImportedItems), '*');
+				window.postMessage(
+					JSON.stringify({
+						payload: mockImportedItems,
+						action: 'getCardBlocks',
+					}),
+					'*'
+				);
 			});
 
 		const items: string[] = ['1', '2', '3'];
@@ -118,7 +133,9 @@ describe('<Importer>', () => {
 			body: [],
 		}).as('fetch');
 
-		cy.mountWithStore(<Importer items={items} />, { reduxStore: store });
+		cy.mountWithStore(<Importer items={items} mode="import" />, {
+			reduxStore: store,
+		});
 
 		cy.get('@boardGetStub').should('be.called');
 
@@ -137,7 +154,7 @@ describe('<Importer>', () => {
 			cy.intercept('POST', '**/api/v3/items/query').as('fetch');
 
 			cy.mountWithStore(
-				<Importer items={[]} queryString={mockQueryString} />
+				<Importer items={[]} mode="import" queryString={mockQueryString} />
 			);
 
 			cy.wait('@fetch')
@@ -149,7 +166,13 @@ describe('<Importer>', () => {
 			cy.stub(window.parent, 'postMessage')
 				.as('boardGetStub')
 				.callsFake(() => {
-					window.postMessage(JSON.stringify(mockImportedItems), '*');
+					window.postMessage(
+						JSON.stringify({
+							payload: mockImportedItems,
+							action: 'getCardBlocks',
+						}),
+						'*'
+					);
 				});
 
 			const mockQueryString = 'item.id IN (1,2,3,4)';
@@ -162,7 +185,7 @@ describe('<Importer>', () => {
 			}).as('fetch');
 
 			cy.mountWithStore(
-				<Importer items={[]} queryString={mockQueryString} />
+				<Importer items={[]} mode="import" queryString={mockQueryString} />
 			);
 
 			cy.get('@boardGetStub').should('be.called');
@@ -176,9 +199,104 @@ describe('<Importer>', () => {
 				.should('equal', expectedQuery);
 		});
 	});
+
+	context('Lines', () => {
+		it('creates lines if mode equals "createLines"', () => {
+			cy.stub(window.parent, 'postMessage').as('postMessageStub');
+
+			cy.mountWithStore(
+				<Importer
+					items={[]}
+					mode="createLines"
+					relationsToCreate={mockRelations}
+					isLoadingRelations={false}
+				/>
+			);
+
+			mockRelations.forEach((relation) => {
+				const linePayload = {
+					sourceBlockId: relation.sourceBlockId,
+					targetBlockId: relation.targetBlockId,
+					relationshipType: relation.type,
+					lineColor: '#000000',
+				};
+				cy.get('@postMessageStub').should(
+					'not.have.been.calledWith',
+					{
+						action: 'createLine',
+						payload: linePayload,
+					},
+					'*'
+				);
+			});
+		});
+
+		it('deletes lines if mode equals "deleteLines"', () => {
+			cy.stub(window.parent, 'postMessage').as('postMessageStub');
+
+			cy.mountWithStore(
+				<Importer
+					items={[]}
+					mode="deleteLines"
+					relationsToDelete={mockLines}
+					isLoadingRelations={false}
+				/>
+			);
+
+			mockLines.forEach((line) => {
+				const linePayload = {
+					lineId: line.id,
+				};
+				cy.get('@postMessageStub').should(
+					'not.have.been.calledWith',
+					{
+						action: 'deleteLine',
+						payload: linePayload,
+					},
+					'*'
+				);
+			});
+		});
+	});
 });
 
-const mockImportedItems = [
-	{ cardBlock: { id: '1' }, codebeamerItemId: 569657 },
-	{ cardBlock: { id: '2' }, codebeamerItemId: 569527 },
+const mockImportedItems: CardBlockToCodebeamerItemMapping[] = [
+	{ cardBlockId: '1', codebeamerItemId: 569657, codebeamerTrackerId: 1 },
+	{ cardBlockId: '1', codebeamerItemId: 569527, codebeamerTrackerId: 1 },
+];
+
+const mockRelations: BlockRelation[] = [
+	{
+		sourceBlockId: '1',
+		targetBlockId: '3',
+		type: RelationshipType.DOWNSTREAM,
+	},
+	{
+		sourceBlockId: '2',
+		targetBlockId: '4',
+		type: RelationshipType.DOWNSTREAM,
+	},
+	{
+		sourceBlockId: '3',
+		targetBlockId: '5',
+		type: RelationshipType.DOWNSTREAM,
+	},
+];
+
+const mockLines: LucidLineData[] = [
+	{
+		id: '1',
+		sourceBlockId: '1',
+		targetBlockId: '3',
+	},
+	{
+		id: '2',
+		sourceBlockId: '2',
+		targetBlockId: '4',
+	},
+	{
+		id: '3',
+		sourceBlockId: '3',
+		targetBlockId: '5',
+	},
 ];